+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 1.0.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Fix broken links generation list.
+  [cekk]
 
 1.0.3 (2018-10-18)
 ------------------